Extract commit_with_signature, sign_buffer, and signatures
