+++ crates/gitbutler-branch-actions/tests/extra/mod.rs
@@ -20,7 +20,7 @@ use gitbutler_branch_actions::{
 };
 use gitbutler_commit::{commit_ext::CommitExt, commit_headers::CommitHeadersV2};
 use gitbutler_reference::{Refname, RemoteRefname};
-use gitbutler_repo::RepositoryExt;
+use gitbutler_repo::committing::RepositoryExt as _;
 use gitbutler_stack::{BranchOwnershipClaims, Target, VirtualBranchesHandle};
 use gitbutler_testsupport::{commit_all, virtual_branches::set_test_target, Case, Suite};
 use pretty_assertions::assert_eq;
