+++ crates/gitbutler-stack/src/stack.rs
@@ -11,9 +11,9 @@ use gitbutler_command_context::CommandContext;
 use gitbutler_commit::commit_ext::CommitExt;
 use gitbutler_id::id::Id;
 use gitbutler_reference::{normalize_branch_name, Refname, RemoteRefname, VirtualRefname};
+use gitbutler_repo::identity::RepositoryExt as _;
 use gitbutler_repo::logging::LogUntil;
 use gitbutler_repo::logging::RepositoryExt as _;
-use gitbutler_repo::RepositoryExt;
 use gix::validate::reference::name_partial;
 use gix_utils::str::decompose;
 use itertools::Itertools;
