+++ crates/gitbutler-testsupport/src/suite.rs
@@ -5,7 +5,7 @@ use std::{
 };
 
 use gitbutler_command_context::CommandContext;
-use gitbutler_repo::RepositoryExt;
+use gitbutler_repo::committing::RepositoryExt as _;
 use tempfile::{tempdir, TempDir};
 
 use crate::test_project::setup_config;
