+++ crates/gitbutler-testsupport/src/testing_repository.rs
@@ -3,7 +3,7 @@ use std::fs;
 use crate::init_opts;
 use gitbutler_commit::commit_headers::CommitHeadersV2;
 use gitbutler_oxidize::git2_to_gix_object_id;
-use gitbutler_repo::RepositoryExt;
+use gitbutler_repo::committing::RepositoryExt as _;
 use gix_testtools::bstr::ByteSlice as _;
 use tempfile::{tempdir, TempDir};
 use uuid::Uuid;
