+++ crates/gitbutler-workspace/src/branch_trees.rs
@@ -4,8 +4,9 @@ use gitbutler_command_context::CommandContext;
 use gitbutler_commit::commit_ext::CommitExt as _;
 use gitbutler_oxidize::{git2_to_gix_object_id, gix_to_git2_oid, GixRepositoryExt};
 use gitbutler_project::access::WorktreeWritePermission;
-use gitbutler_repo::rebase::cherry_rebase_group;
+use gitbutler_repo::committing::RepositoryExt as _;
 use gitbutler_repo::RepositoryExt as _;
+use gitbutler_repo::{identity::RepositoryExt as _, rebase::cherry_rebase_group};
 use gitbutler_stack::{Stack, VirtualBranchesHandle};
 use tracing::instrument;
